% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{clin_nowcast}
\alias{clin_nowcast}
\title{Nowcast clinical laboratory percent positive flu data}
\usage{
clin_nowcast(clin, weeks_to_replace = 1)
}
\arguments{
\item{clin}{Data prepared with \code{\link[=get_cdc_clin]{get_cdc_clin()}}}

\item{weeks_to_replace}{Number of retrospective weeks to replace with nowcast; default is 1}
}
\value{
A tibble formatted the same as that returned with \code{get_cdc_clin()} but where the n most recent weeks (n="weeks_to_replace") have been nowcasted.
}
\description{
This function provides a naive nowcasting method for clinical laboratory percent positive flu data. The methodology simply averages the last 4 weeks of available data and uses this average as the value for the number of weeks specified to replace. This is useful given that there is reporting lag in the NREVSS clinical laboratory percent positive flu data.
}
