% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{density_probs}
\alias{density_probs}
\title{Title}
\usage{
density_probs(df, n_horizons = 5, ...)
}
\arguments{
\item{df}{Data}

\item{n_horizons}{Number of horizons ahead}

\item{...}{Additional arguments passed to distfromq}
}
\value{
data frame
}
\description{
Title
}
