% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.mcga}
\alias{.mcga}
\title{Make clean column names}
\usage{
.mcga(tbl)
}
\arguments{
\item{tbl}{Input tibble with columns to rename}
}
\value{
Tibble with clean column names
}
\description{
This helper function is used
}
