% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wis.R
\name{find_quantile_match}
\alias{find_quantile_match}
\title{Find quantile that matches value}
\usage{
find_quantile_match(quantiles, val_to_match, tol = 1e-08)
}
\arguments{
\item{quantiles}{vector of quantile}

\item{val_to_match}{quantile to match}

\item{tol}{tolerance}
}
\value{
a logical vector indicating \code{TRUE} for the \code{val_to_match} in \code{quantiles}.
}
\description{
Unexported utility function to find quantile that matches a value. See \link{weighted_interval_score}.
}
