% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fiphde-package.R
\docType{package}
\name{fiphde-package}
\alias{fiphde}
\alias{fiphde-package}
\title{fiphde: Forecasting Influenza in Support of Public Health Decision Making}
\description{
Miscellaneous functions for retrieving data, creating and evaluating time series 
    forecasting models for influenza-like illness (ILI) and influenza hospitalizations in 
    the United States.
}
\author{
\strong{Maintainer}: VP Nagraj \email{pnagraj@signaturescience.com} (\href{https://orcid.org/0000-0003-0060-566X}{ORCID})

Authors:
\itemize{
  \item Shakeel Jessa \email{sjessa@signaturescience.com}
  \item Chris Hulme-Lowe \email{mscholz@signaturescience.com}
  \item Stephen Turner \email{sturner@signaturescience.com} (\href{https://orcid.org/0000-0001-9140-9028}{ORCID})
}

}
\keyword{internal}
