% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny.R
\name{fiphde_launcher}
\alias{fiphde_launcher}
\title{FIPHDE explorer app launcher}
\usage{
fiphde_launcher(.data, submission_dir, app_dir = NULL, ...)
}
\arguments{
\item{.data}{Tibble with historical data for trend leading up to forecast}

\item{submission_dir}{Full path to directory of submission files containing forecast submissions to explore}

\item{app_dir}{Full path to directory of explorer app; default is \code{NULL} and app directory will be resolved from \code{system.file("app", package="fiphde")}}

\item{...}{Additional arguments to be passed to \link[shiny:runApp]{shiny::runApp}}
}
\value{
This function starts a shiny app. On exit it removes objects (see ".data" and "submission_dir") that are temporarily attached and used by the app session.
}
\description{
The explorer app allows a user to view plots of forecasts, inspect tabular output of submission files, and download subsets of forecast submission data. The app includes an interface to interactively select locations to include in the plots, table, and download. This function wraps \code{shiny::runApp} and accepts arguments for the data against which the forecasts should be plotted, as well as the directory containing submission files, both of which are temporarily attached to the global environment for use during the app session. Additional arguments passed to \code{...} will be inherited by \link[shiny:runApp]{shiny::runApp}.
}
