% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast_categorical}
\alias{forecast_categorical}
\title{Forecast categorical targets}
\usage{
forecast_categorical(
  .forecast,
  .observed,
  method = "density",
  format = "hubverse",
  horizon = 4
)
}
\arguments{
\item{.forecast}{A tibble with "submission-ready" probabilistic flu hospitalization forecast data (i.e., tibble contained in list element returned from \link{format_for_submission})}

\item{.observed}{A tibble with observed flu admission data (i.e., tibble output from \link{prep_hdgov_hosp})}

\item{method}{The categorical forecasting method to use; must be one of \code{"density"} or \code{"interpolation"}; default is \code{"density"}}

\item{format}{The submission format to be used; must be one of \code{"hubverse"} or \code{"legacy"}; default is \code{"hubverse"}}

\item{horizon}{The number of horizons ahead to forecast; must be one of \code{4} or \code{5}; default is \code{4}}
}
\value{
A \code{tibble} with formatted categorical forecasts.

If format is \code{"hubverse"} the tibble will have the following columns:
\itemize{
\item \strong{reference_date}: Date of reference for forecast submission
\item \strong{horizon}: Horizon for the given forecast
\item \strong{target}: Name of forecasted target
\item \strong{target_end_date}: Last date of the forecasted target (e.g., Saturday of the given epidemiological week)
\item \strong{location}: Name or geographic identifier (e.g., FIPS code) for location for the given forecast
\item \strong{output_type}: Type of forecasted value (e.g., "quantile")
\item \strong{output_type_id}: The quantile for the forecasted value if output_type is "quantile"
\item \strong{value}: The forecasted value
}

If format is \code{"legacy"} the tibble will have the following columns:
\itemize{
\item \strong{forecast_date}: Date of forecast
\item \strong{target}: Horizon and name of forecasted target
\item \strong{target_end_date}: Last date of the forecasted target (e.g., Saturday of the given epidemiological week)
\item \strong{location}: Name or geographic identifier (e.g., FIPS code) for location for the given forecast
\item \strong{type}: One of either "point" or "quantile" for the forecasted value
\item \strong{quantile}: The quantile for the forecasted value; \code{NA} if "type" is \code{"point"}
\item \strong{value}: The forecasted value
}
}
\description{
This function takes probabilistic flu hospitalization forecast input and converts the forecasted values for each location to a categorical "change" indicator. The criteria for each level ("large decrease", "decrease", "stable", "increase", "large increase") was defined by the CDC (see link in references). The algorithm evaluates absolute changes in counts and rates (per 100k individuals) for the most recently observed week and a 2 week ahead forecasted horizon. This procedure runs independently for each location, and results in a formatted tabular output that includes each possible level and its corresponding probability of being observed (calculated from probabilistic quantiles) for every location.
}
\examples{
\dontrun{
# Retrieve hospitalization data
h_raw <- get_hdgov_hosp(limitcols=TRUE)
# Prepare and summarize hospitalization data to weekly resolution
prepped_hosp <- prep_hdgov_hosp(h_raw)
# Create a keyed time series tibble with only locations of interest
prepped_tsibble <- make_tsibble(prepped_hosp,
                                     epiyear = epiyear,
                                     epiweek=epiweek,
                                     key=location)

# Run with default constrained ARIMA, nonseasonal ETS, no NNETAR
hosp_fitfor <- ts_fit_forecast(prepped_tsibble,
                               horizon=4L,
                               outcome="flu.admits",
                               covariates=TRUE)
# Prepare forecast for quantile submission format
forc <- format_for_submission(hosp_fitfor$tsfor, method = "ts", format = "legacy")
# Run categorical summary of quantiles for the time series ensemble
forecast_categorical(forc$ensemble, prepped_hosp, method = "interpolation", format = "legacy")
}
}
\references{
\url{https://github.com/cdcepi/Flusight-forecast-data/blob/master/data-experimental/README.md}
}
