% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{get_cdc_clin}
\alias{get_cdc_clin}
\title{Retrieve and prep clinical laboratory percent positive flu data}
\usage{
get_cdc_clin(region = "both", years = NULL)
}
\arguments{
\item{region}{Either "state", "national", or "both". Defaults to "both" to return state and national data combined.}

\item{years}{A vector of years to retrieve data for. CDC has data going back to 1997. Default value (\code{NULL}) retrieves \strong{all} years.}
}
\value{
Tibble with prepared data pulled from NREVSS, including columns for total number of positives, percent positive, epiweek, epiyear
}
\description{
This function returns weekly state and/or national clinical laboratory percent positivity data from the NREVSS reporting instrument.
}
