% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{get_cdc_hosp}
\alias{get_cdc_hosp}
\title{Get hospitalization data from CDC FluView}
\usage{
get_cdc_hosp(years = NULL)
}
\arguments{
\item{years}{A vector of years to retrieve data for (i.e. 2014 for CDC flu season 2014-2015). CDC has data going back to 2009 and up until the \emph{previous} flu season. Default value (\code{NULL}) retrieves \strong{all} years.}
}
\value{
A tibble
}
\description{
Get hospitalization data from CDC FluView. See \link[cdcfluview:hospitalizations]{cdcfluview::hospitalizations}.
}
\examples{
\dontrun{
get_cdc_hosp(years=2019)
}
}
\references{
cdcfluview documentation: \url{https://hrbrmstr.github.io/cdcfluview/index.html#retrieve-ilinet-surveillance-data}.
}
