% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{get_cdc_vax}
\alias{get_cdc_vax}
\title{Get vaccination data}
\usage{
get_cdc_vax(
  endpoint = "https://data.cdc.gov/api/views/k87d-gv3u/rows.csv",
  app_token = Sys.getenv("HEALTHDATA_APP_TOKEN")
)
}
\arguments{
\item{endpoint}{URL to cdc.gov endpoint. See references.}

\item{app_token}{App token from healthdata.gov. If it's \code{NULL} you might get rate limited. Add an entry to your \verb{~/.Renviron} with \code{HEALTHDATA_APP_TOKEN="tokenhere"} that you got from \url{https://healthdata.gov/profile/edit/developer_settings}.}
}
\value{
A tibble
}
\description{
Get vaccination data from cdc.gov endpoint.
}
\references{
API documentation: \url{https://dev.socrata.com/foundry/data.cdc.gov/k87d-gv3u}.
}
