% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{get_hdgov_hosp}
\alias{get_hdgov_hosp}
\title{Get hospitalization data}
\usage{
get_hdgov_hosp(
  endpoint = "https://healthdata.gov/resource/g62h-syeh.json",
  state = NULL,
  limitrows = NULL,
  mindate = NULL,
  maxdate = NULL,
  limitcols = FALSE,
  app_token = Sys.getenv("HEALTHDATA_APP_TOKEN")
)
}
\arguments{
\item{endpoint}{URL to healthdata.gov endpoint (see references)..}

\item{state}{A two-letter state abbreviation.}

\item{limitrows}{Limit API query to at most this number of results. Default is \code{NULL} (no limit).}

\item{mindate}{Minimum date of results returned (ISO 8601 format: YYYY-MM-DD). See examples. Note that after retrieving the data, dates will be shifted back one day because the admissions data is \emph{previous} day's flu/covid admissions.}

\item{maxdate}{Maximum date of results returned (ISO 8601 format: YYYY-MM-DD). See examples. Note that after retrieving the data, dates will be shifted back one day because the admissions data is \emph{previous} day's flu/covid admissions.}

\item{limitcols}{Limit the columns returned to the subjectively defined important ones?}

\item{app_token}{App token from healthdata.gov. If \code{NULL} you might get rate limited. Add an entry to your \verb{~/.Renviron} with \code{HEALTHDATA_APP_TOKEN="tokenhere"} that you got from \url{https://healthdata.gov/profile/edit/developer_settings}.}
}
\value{
A tibble
}
\description{
Retrieves hospitalization data from the healthdata.gov endpoint with optional filtering on fields, and return the results into a nice tibble.
}
\examples{
\dontrun{
get_hdgov_hosp(mindate="2021-11-01", limitrows=10, limitcols=TRUE)
get_hdgov_hosp(mindate="2021-11-01", limitrows=10, limitcols=FALSE)
get_hdgov_hosp(state="VA")
get_hdgov_hosp(state="VA", mindate="2021-10-01")
get_hdgov_hosp(state="VA", mindate="2021-10-01", maxdate="2021-10-31")
get_hdgov_hosp(state="VA", mindate="2021-10-01", maxdate="2021-11-21", limitrows=5)
}
}
\references{
API documentation: \url{http://dev.socrata.com/foundry/healthdata.gov/g62h-syeh}.
}
