% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{get_nhsn_weekly}
\alias{get_nhsn_weekly}
\title{Retrieve weekly NHSN flu hospitalization data}
\usage{
get_nhsn_weekly(endpoint = "https://data.cdc.gov/api/views/mpgq-jmmr/rows.csv")
}
\arguments{
\item{endpoint}{URL to data.cdc.gov endpoint; default is \code{"https://data.cdc.gov/api/views/mpgq-jmmr/rows.csv"}}
}
\value{
A \code{tibble} with the following columns:
\itemize{
\item \strong{abbreviation}: Abbreviation of the state or US aggregate
\item \strong{week_end}: End date for the epiweek/epiyear being reproted
\item \strong{flu.admits}: Count of incident flu cases among hospitalized patients
\item \strong{flu.admits.cov}: Coverage (number of hospitals reporting) for incident flu cases
\item \strong{flu.admits.cov.perc}: Coverage (percentage of hospitals reporting) for incident flu cases
}
}
\description{
This function retrieves weekly aggregated NHSN hospital respiratory data API. The function was written to use the default API endpoint (see description of "endpoint" argument and link in references). Note one the available endpoints includes data flagged as "preliminary". All reported weekly aggregates include the number of facilities reporting. In the weeks between April 28, 2024 and November 02, 2024 the NHSN flu hospitalization signal was not required to be reported.
}
\references{
\url{https://data.cdc.gov/Public-Health-Surveillance/Weekly-Hospital-Respiratory-Data-HRD-Metrics-by-Ju/mpgq-jmmr/about_data}

\url{https://data.cdc.gov/Public-Health-Surveillance/Weekly-Hospital-Respiratory-Data-HRD-Metrics-by-Ju/ua7e-t2fy/about_data}
}
