% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{get_nowcast_ili}
\alias{get_nowcast_ili}
\title{Retrieve ILI nowcast}
\usage{
get_nowcast_ili(
  epiyearweeks = NULL,
  dates = lubridate::today() - c(14, 7),
  state = NULL,
  boundatzero = TRUE
)
}
\arguments{
\item{epiyearweeks}{A vector of epiyear-epiweeks to retrieve data for, e.g., 202150, 202151, etc. Exclusive with dates}

\item{dates}{A vector of dates to retrieve data for, e.g., ""2021-12-12" or "2021-12-19". Exclusive with epiyearweek. Defaults to two weeks prior.}

\item{state}{A vector of states to retrieve (two-letter abbreviation). Default \code{NULL} retrieves all states, national, and hhs regions. See examples.}

\item{boundatzero}{Logical as to whether or not the values should be truncated at 0 (i.e., non-negative); default is \code{TRUE}}
}
\value{
A \code{tibble}
}
\description{
This function pulls the ILI nowcast from CMU Delphi's ILI Nearby API. Observed ILINet data is typically reported with a lag, and the ILI nowcast can be used to augment the ILI data stream.
}
\examples{
\dontrun{
# Defaults to the previous two weeks for all states
get_nowcast_ili()

# Otherwise specify one or the other, not both
get_nowcast_ili(epiyearweeks=c("202150", "202151"), dates=NULL)
get_nowcast_ili(epiyearweeks=NULL, dates=c("2021-12-12", "2021-12-19"))

# Get just one state for the last years worth of data (back 52 weeks to 1 week)
get_nowcast_ili(epiyearweeks=NULL,
                dates=lubridate::today()-seq(52*7, 7, -7),
                state="FL")
}
}
\references{
\url{https://delphi.cmu.edu/nowcast/}
}
