% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{get_respnet}
\alias{get_respnet}
\title{Retrieve data from RESP-NET}
\usage{
get_respnet(network = "FluSurv-NET")
}
\arguments{
\item{network}{The name of the RESP-NET network to query; must be one of \code{"FluSurv-NET"}, \code{"COVID-NET"}, \code{"RSV-NET"}, or \code{"Combined"}; default is \code{"FluSurv-NET"}}
}
\description{
This unexported helper function retrieves data from the RESP-NET API for respiratory hospitalizations. The data retrieved can be filtered to specific network(s) reported. The function is used internally by \link{get_cdc_hosp}.
}
\references{
\itemize{
\item \href{https://data.cdc.gov/Public-Health-Surveillance/Rates-of-Laboratory-Confirmed-RSV-COVID-19-and-Flu/kvib-3txy/about_data}{RESP-NET Data Portal}
\item \href{https://www.cdc.gov/resp-net/dashboard/}{RESP-NET Dashboard}
}
}
