% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{glm_quibble}
\alias{glm_quibble}
\title{Get quantiles from prediction intervals}
\usage{
glm_quibble(
  fit,
  new_data,
  alpha = c(0.01, 0.025, seq(0.05, 0.45, by = 0.05)) * 2
)
}
\arguments{
\item{fit}{Fitted model object from \link[fiphde]{glm_fit}}

\item{new_data}{Tibble with new data on which the \link[trending]{predict.trending_model_fit} method should run}

\item{alpha}{Vector specifying the threshold(s) to be used for prediction intervals; alpha of \code{0.05} would correspond to 95\% PI; default is \code{c(0.01, 0.025, seq(0.05, 0.45, by = 0.05)) * 2} to range of intervals}
}
\value{
A tibble with predicted values at each quantile (lower and upper bound for each value of "alpha")
}
\description{
This function runs the \link[trending]{predict.trending_model_fit} method on a fitted model at specified values of "alpha" in order to create a range of prediction intervals. The processing also includes steps to convert the alpha to corresponding quantile values at upper and lower bounds.
}
\examples{
\dontrun{
##FIXME
}
}
