% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{glm_wrap}
\alias{glm_wrap}
\title{Run glm modeling and forecasting}
\usage{
glm_wrap(
  .data,
  .models,
  new_covariates = NULL,
  horizon = 4,
  alpha = c(0.01, 0.025, seq(0.05, 0.45, by = 0.05)) * 2
)
}
\arguments{
\item{.data}{Data including all explanatory and outcome variables needed for modeling}

\item{.models}{List of models defined as \link[trending]{trending_model} objects}

\item{new_covariates}{Tibble with one column per covariate, and n rows for n horizons being forecasted}

\item{horizon}{Number of weeks ahead for forecasting}

\item{alpha}{Vector specifying the threshold(s) to be used for prediction intervals; alpha of \code{0.05} would correspond to 95\% PI; default is \code{c(0.01, 0.025, seq(0.05, 0.45, by = 0.05)) * 2} to range of intervals}
}
\value{
Named list with two elements:
\itemize{
\item model: Output from \link[fiphde]{glm_fit} with selected model fit
\item forecasts: Output from \link[fiphde]{glm_forecast} with forecasts from each horizon combined as a single tibble
}
}
\description{
This is a wrapper function that pipelines influenza hospitalization modeling (\link[fiphde]{glm_fit}) and forecasting (\link[fiphde]{glm_forecast}).
}
