% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit.R
\name{hubverse_format}
\alias{hubverse_format}
\title{Hubverse formatting}
\usage{
hubverse_format(dat, horizon_shift = 1)
}
\arguments{
\item{dat}{Forecast prepped in "legacy" format}

\item{horizon_shift}{Number of horizons to shift backwards to align with reference date; default is \code{1}}
}
\value{
Formatted \code{tibble}
}
\description{
This unexported helper is used internally inside in \link{format_for_submission}.It specifically updates formatting for Hubverse guidelines.
}
\references{
\url{https://github.com/signaturescience/FluSight-forecast-hub/tree/main/model-output#forecast-file-format}

\url{https://github.com/cdcepi/Flusight-forecast-data/blob/master/data-forecasts/README.md}
}
