% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{ilinet}
\alias{ilinet}
\title{Retrieve ILINet surveillance data}
\usage{
ilinet(region = c("national", "hhs", "census", "state"), years = NULL)
}
\arguments{
\item{region}{One of "\code{national}", "\code{hhs}", "\code{census}", or "\code{state}"}

\item{years}{A vector of years to retrieve data for (i.e. \code{2014} for CDC
flu season 2014-2015). CDC has data for this API going back to 1997.
Default value (\code{NULL}) means retrieve \strong{all} years. NOTE: if you
happen to specify a 2-digit season value (i.e. \code{57} == 2017-2018)
the function is smart enough to retrieve by season ID vs convert that
to a year.}
}
\description{
Adapted from \code{cdcfluview::ilinet}.

This unexported helper function retrieves current and historical ILINet surveillance data for
the identified region via the CDC FluView API. The function is used internally in \link{get_cdc_ili}. Data returned include weighted and unweighted ILI percentage, as well as age-specific ILI outpatient visit counts for each location / epidemiological week.
}
\references{
\itemize{
\item \href{https://github.dev/hrbrmstr/cdcfluview}{cdcfluview package}
\item \href{https://gis.cdc.gov/grasp/fluview/fluportaldashboard.html}{CDC FluView Portal}
}
}
