% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mmwr_week_to_date}
\alias{mmwr_week_to_date}
\title{Convert MMWR format to date}
\usage{
mmwr_week_to_date(year, week, day = NULL)
}
\arguments{
\item{year}{Vector of epidemiological year(s); must be same length as "week" and "day" (unless "day" is \code{NULL})}

\item{week}{Vector of epidemiological week(s); must be same length as "year" and "day" (unless "day" is \code{NULL})}

\item{day}{Vector of day(s); must be same length as "week" and "year" (unless set to is \code{NULL}); default is \code{NULL} and the day returned will be the first day of the epidemiological week (i.e., Sunday)}
}
\value{
Vector of date objects as with as many elements as input year(s), week(s), day(s)
}
\description{
Adapted from \code{cdcfluview::mmwr_week_to_date}.
}
\details{
This function transforms MMWR epidemiological year+week (or year+week+day) to a date object. This was implemented based on the \code{cdcfluview::mmwr_week_to_date} function, which adapted similar functionality from the \code{MMWRweek} package.
}
\examples{
mmwr_week_to_date(2020,1)
mmwr_week_to_date(2020,1,5)
mmwr_week_to_date(c(2020,2021,2022),c(1,2,8), c(1,1,7))
}
\references{
\itemize{
\item \href{https://github.dev/hrbrmstr/cdcfluview}{cdcfluview package}
}
}
