% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mnz}
\alias{mnz}
\title{Minimum non-zero}
\usage{
mnz(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
The minimum non-zero positive value from \code{x}
}
\description{
Helper function to get the minimum non-zero positive value from a vector. Used internally in \link{mnz_replace}
}
\examples{
x <- c(.1, 0, -.2, NA, .3, .4, .0001, -.3, NA, 999)
x
mnz(x)
}
