% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mnz_replace}
\alias{mnz_replace}
\title{Minimum non-zero replacement}
\usage{
mnz_replace(x)
}
\arguments{
\item{x}{A numeric vector}
}
\value{
A vector of the same length with negatives and zeros replaced with the minimum nonzero value of that vector.
}
\description{
Replace zeros and negative values with the minimum non-zero positive value from a vector.
}
\examples{
x <- c(.1, 0, -.2, NA, .3, .4, .0001, -.3, NA, 999)
x
mnz(x)
mnz_replace(x)
tibble::tibble(x) \%>\% dplyr::mutate(x2=mnz_replace(x))
}
