% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_forc}
\alias{plot_forc}
\title{Plot forecasts against observed data}
\usage{
plot_forc(.forecasts, .train, .test)
}
\arguments{
\item{.forecasts}{Tibble with prepped forecasts}

\item{.train}{Tibble with data used for modeling}

\item{.test}{Tibble with observed data held out from modeling}
}
\value{
A \code{ggplot2} plot object
}
\description{
This helper function creates a plot to visualize the forecasted point estimates (and 95\% prediction interval) alongside the observed data.
}
