% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prep.R
\name{prep_hdgov_hosp}
\alias{prep_hdgov_hosp}
\title{Prep hospitalization data}
\usage{
prep_hdgov_hosp(
  h_raw,
  statesonly = TRUE,
  trim = list(epiyear = 2020, epiweek = 43),
  remove_incomplete = TRUE,
  min_per_week = 1
)
}
\arguments{
\item{h_raw}{Raw hospitalization data from \link{get_hdgov_hosp}}

\item{statesonly}{Limit to US+DC+States only (i.e., drop territories)? Defaults to \code{TRUE}.}

\item{trim}{list with epiyear and epiweek to trim on. Defaults to October 25, 2020 (2020:43). There isn't any data before this.}

\item{remove_incomplete}{Remove the last week if incomplete? Defaults to \code{TRUE}.}

\item{min_per_week}{The minimum number of flu.admits per week needed to retain that state. Default removes states with less than 1 flu admission per week over the last 30 days.}
}
\value{
A tibble; hospitalization data ready for downstream forecasting.
}
\description{
Prep healthdata.gov hospitalization data retrieved using \link{get_hdgov_hosp} for downstream forecasting. Limits to states only, trims to where you have data, removes the incomplete week, and removes locations with little reporting over the last month.
}
\examples{
\dontrun{
h_raw <- get_hdgov_hosp(limitcols=TRUE)
prep_hdgov_hosp(h_raw)
}
}
