% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{round_preserve}
\alias{round_preserve}
\title{Round and preserve vector}
\usage{
round_preserve(x, digits = 0)
}
\arguments{
\item{x}{Numeric vector with values to round}

\item{digits}{The number of digits to use in precision; defalut is \code{0}}
}
\value{
Vector of same length as "x" with values rounded
}
\description{
This unexported helper is used to ensure that categorical forecasts are rounded to sum to 1.
}
