% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wis.R
\name{score_func_param_checker}
\alias{score_func_param_checker}
\title{Score function parameter checker}
\usage{
score_func_param_checker(quantiles, values, actual_value, id = "")
}
\arguments{
\item{quantiles}{Vector of forecasted quantiles}

\item{values}{Vector of forecasted values}

\item{actual_value}{Actual value}

\item{id}{Name of the score function, e.g. "weighted_interval_score"}
}
\value{
Nothing. Called for side effects.
}
\description{
This function is adapted from the \code{evalcast::score_func_param_checker}.

Unexported utility function to check parameters of a score function. Used in \link{weighted_interval_score}.
}
\references{
\url{https://github.com/cmu-delphi/covidcast/tree/main/R-packages/evalcast}
}
