% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{state_replace_ili_nowcast_all}
\alias{state_replace_ili_nowcast_all}
\title{Replace ILINet data with nowcast entirely for a state}
\usage{
state_replace_ili_nowcast_all(
  ilidat,
  state,
  impute = TRUE,
  fallback = TRUE,
  ...
)
}
\arguments{
\item{ilidat}{ILI data retrieved via \link{get_cdc_ili}}

\item{state}{Two-letter state abbreviation to replace completely}

\item{impute}{Logical as to whether or not to try to mean impute missing values using the immediately preceding and following values; default is \code{TRUE}}

\item{fallback}{Logical as to whether or not to fall back to pseudo nowcast (average of last 4 ILI weeks in the given location) if nowcast data is unavailable; default is \code{TRUE}}

\item{...}{Other arguments passed to \link{get_nowcast_ili} (e.g., \code{boundatzero}, which is \code{TRUE} by default)}
}
\value{
A \code{tibble} with the following columns:
\itemize{
\item \strong{location}: FIPS code for the location
\item \strong{region_type}: The type of location
\item \strong{abbreviation}: Abbreviation for the location
\item \strong{region}: Name of the region
\item \strong{epiyear}: Year of reporting (in epidemiological week calendar)
\item \strong{epiweek}: Week of reporting (in epidemiological week calendar)
\item \strong{week_start}: Date of beginning (Sunday) of the given epidemiological week
\item \strong{weighted_ili}: Population-weighted percentage of ILI outpatient visits
}
}
\description{
This function replaces the weighted ILI retrieved from \link{get_cdc_ili} with nowcast data for each of the locations in the original data. The function will first attempt to use ILI Nearby nowcasts pulled using \link{get_nowcast_ili}. If the ILI Nearby nowcasts are unavailable, the function will optionally fallback to a pseudo nowcast method that averages the observed ILI for the 4 most recent weeks. Unlike \link{replace_ili_nowcast}, the ILI will be replaced with nowcast for \emph{all} dates for a specified location. This is useful for getting data for states where most or all ILI data is missing (e.g., Florida).

Note that this only replaces weighted ILI in the specified state where the value is \code{NA}. \emph{Most} ILI data from Florida is missing, but not all.
}
\examples{
\dontrun{

ilidat <- get_cdc_ili(years=2020)

ilidat <-
  ilidat \%>\%
  dplyr::filter(location=="US" | abbreviation=="VA" | abbreviation=="FL") \%>\%
  dplyr::group_by(location) \%>\%
  dplyr::slice_max(week_start, n=4) \%>\%
  dplyr::select(location:weighted_ili) \%>\%
  dplyr::arrange(location, epiyear, epiweek)
ilidat
state_replace_ili_nowcast_all(ilidat, state="FL")
# Example with Florida, which has a negative value for nowcasted ILI
ilidat <- get_cdc_ili(years=2019)
ilidat <- ilidat \%>\%
  dplyr::filter(location=="US" | abbreviation=="VA" | abbreviation=="FL") \%>\%
  dplyr::filter(epiyear==2020 & epiweek \%in\% c(20, 21, 22)) \%>\%
  dplyr::select(location:weighted_ili) \%>\%
  dplyr::arrange(location, epiyear, epiweek)
ilidat
# defaults to bound at zero
state_replace_ili_nowcast_all(ilidat, state="FL")
# show results when you don't bound at zero
state_replace_ili_nowcast_all(ilidat, state="FL", boundatzero=FALSE)
# example with missing data in florida
ilidat <- get_cdc_ili(region=c("national","state"), years=2019:lubridate::year(lubridate::today()))
ilidat <- ilidat \%>\%
  dplyr::filter(abbreviation=="FL") \%>\%
  dplyr::filter(week_start>="2020-12-13" & week_start<="2021-01-10")
ilidat
state_replace_ili_nowcast_all(ilidat, state="FL")
state_replace_ili_nowcast_all(ilidat, state="FL", impute=FALSE)
}
}
