% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{surveillance_areas}
\alias{surveillance_areas}
\title{Retrieve a list of valid sub-regions for each surveillance area.}
\usage{
surveillance_areas()
}
\description{
Retrieve a list of valid sub-regions for each surveillance area.
}
\examples{
\dontrun{
surveillance_areas()
}
}
