% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{this_saturday}
\alias{this_saturday}
\title{Get Saturday}
\usage{
this_saturday()
}
\value{
Date for the Saturday of the current week.
}
\description{
This function is a helper to get the date for the Saturday of the current week. The function determines the current week based on epidemiological week orientation (i.e., week begins with Sunday).
}
\examples{
this_saturday()
}
