% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{to_num}
\alias{to_num}
\title{Clean numeric values}
\usage{
to_num(x)
}
\arguments{
\item{x}{Input character vector for which special characters should be stripped and converted}
}
\value{
Numeric vector
}
\description{
This unexported helper is used in the \link{ilinet} function to strip special characters and empty space and convert a character vector to numeric.
}
\examples{
x <- c(",1", 4, "11", "19  ", "25>")
fiphde:::to_num(x)
}
