% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{ts_format_for_submission}
\alias{ts_format_for_submission}
\title{Format time series forecast}
\usage{
ts_format_for_submission(tsfor)
}
\arguments{
\item{tsfor}{The forecast from \link{ts_fit_forecast}.}
}
\value{
A named list of tibbles, one for each model, formatted for submission.
}
\description{
Format time series forecast for submission.
}
\details{
Uses quantiles \code{c(0.01, 0.025, seq(0.05, 0.95, by = 0.05), 0.975, 0.99)} in the built-in \code{fiphde:::q}, using an accessory table \code{fiphde:::quidk}. See \code{data-raw/generate-sysdata.R} for details.
}
\examples{
\dontrun{
# Get raw data from healthdata.gov
h_raw <- get_hdgov_hosp(limitcols=TRUE)
## save(h_raw, file="~/Downloads/h_raw.rd")
## load(file="~/Downloads/h_raw.rd")

# Prep, and make a tsibble
prepped_hosp <- prep_hdgov_hosp(h_raw, statesonly=TRUE)
prepped_hosp_tsibble <- make_tsibble(prepped_hosp,
                                     epiyear = epiyear,
                                     epiweek=epiweek,
                                     key=location)
# Limit to only Virginia and US
prepped_hosp_tsibble <-
  prepped_hosp_tsibble \%>\%
  dplyr::filter(location \%in\% c("US", "51"))

# Fit a model
hosp_fitfor <- ts_fit_forecast(prepped_hosp_tsibble,
                               horizon=4L,
                               outcome="flu.admits",
                               constrained=TRUE,
                               covariates=c("hosp_rank", "ili_rank"))

# format for submission
formatted_list <- ts_format_for_submission(hosp_fitfor$tsfor)
formatted_list
}
}
\references{
\url{https://github.com/cdcepi/Flusight-forecast-data/blob/master/data-forecasts/README.md}
}
