% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wis.R
\name{weighted_interval_score}
\alias{weighted_interval_score}
\title{Compute weighted interval score}
\usage{
weighted_interval_score(quantile, value, actual_value)
}
\arguments{
\item{quantile}{Vector of forecasted quantiles}

\item{value}{Vector of forecasted values}

\item{actual_value}{Actual value}
}
\description{
This function is adapted from the \code{evalcast::weighted_interval_score}.

Computes weighted interval score (WIS), a well-known quantile-based
approximation of the commonly-used continuous ranked probability score
(CRPS). WIS is a proper score, and can be thought of as a distributional
generalization of absolute error. For example, see \href{https://arxiv.org/abs/2005.12881}{Bracher et al. (2020)} for discussion in the context
of COVID-19 forecasting.
}
\references{
\url{https://github.com/cmu-delphi/covidcast/tree/main/R-packages/evalcast}
}
