% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{who_nrevss}
\alias{who_nrevss}
\title{WHO/NREVSS surveillance data}
\usage{
who_nrevss(region = c("national", "hhs", "census", "state"), years = NULL)
}
\arguments{
\item{region}{one of "\code{national}", "\code{hhs}", "\code{census}", or "\code{state}"}

\item{years}{a vector of years to retrieve data for (i.e. \code{2014} for CDC
flu season 2014-2015). CDC has data for this API going back to 1997.
Default value (\code{NULL}) means retrieve \strong{all} years. NOTE: if you
happen to specify a 2-digit season value (i.e. \code{57} == 2017-2018)
the function is smart enough to retrieve by season ID vs convert that
to a year.}
}
\description{
Adapted from cdcfluview::who_nrevss.
}
\examples{
\dontrun{
national_who <- who_nrevss("national")
hhs_who <- who_nrevss("hhs")
census_who <- who_nrevss("census")
state_who <- who_nrevss("state")
}
}
\references{
\itemize{
\item \href{https://github.dev/hrbrmstr/cdcfluview}{cdcfluview package}
}
}
