% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{wis_score}
\alias{wis_score}
\title{Calculate WIS score}
\usage{
wis_score(.forecasts, .test)
}
\arguments{
\item{.forecasts}{Tibble with prepped foreacsts}

\item{.test}{Tibble with test data including observed value for flu admissions stored in "flu.admits" column}
}
\value{
Tibble with the WIS for each combination of epiweek and epiyear
}
\description{
Helper function to calculate weighted interval score (WIS) for prepped forecasts
}
