% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.mcga}
\alias{.mcga}
\title{Make clean column names}
\usage{
.mcga(tbl)
}
\arguments{
\item{tbl}{Input \code{tibble} with columns to rename}
}
\value{
A \code{tibble} with clean column names
}
\description{
This helper is used in \link{ilinet} and \link{who_nrevss} functions to clean column names of values returned from the APIs.
}
