% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast_categorical}
\alias{forecast_categorical}
\title{Forecast categorical targets}
\usage{
forecast_categorical(.forecast, .observed)
}
\arguments{
\item{.forecast}{A tibble with "submission-ready" probabilistic flu hospitalization forecast data (i.e., tibble contained in list element returned from \link{format_for_submission})}

\item{.observed}{A tibble with observed flu admission data (i.e., tibble output from \link{prep_hdgov_hosp})}
}
\value{
Tibble with formatted categorical forecasts that includes the following columns:
\itemize{
\item \strong{forecast_date}: Date of forecast
\item \strong{target}: Name of target forecasted; fixed at "2 wk flu hosp rate change"
\item \strong{location}: FIPS code for the location
\item \strong{type}: The type of forecast output; fixed at "category"
\item \strong{type_id}: Categorical label; one of "large decrease", "decrease", "stable", "increase", "large increase"
\item \strong{value}: Probability of observing the given "type_id" at the given "location"
}
}
\description{
This function takes probabilistic flu hospitalization forecast input and converts the forecasted values for each location to a categorical "change" indicator. The criteria for each level ("large decrease", "decrease", "stable", "increase", "large increase") was defined by the CDC (see link in references). The algorithm evaluates absolute changes in counts and rates (per 100k individuals) for the most recently observed week and a 2 week ahead forecasted horizon. This procedure runs independently for each location, and results in a formatted tabular output that includes each possible level and its corresponding probability of being observed (calculated from probabilistic quantiles) for every location.
}
\examples{
\dontrun{
h_raw <- get_hdgov_hosp(limitcols=TRUE)
prepped_hosp <- prep_hdgov_hosp(h_raw)
prepped_tsibble <- make_tsibble(prepped_hosp,
                                     epiyear = epiyear,
                                     epiweek=epiweek,
                                     key=location)

# Run with default constrained ARIMA, nonseasonal ETS, no NNETAR
hosp_fitfor <- ts_fit_forecast(prepped_tsibble,
                               horizon=4L,
                               outcome="flu.admits",
                               covariates=c("hosp_rank", "ili_rank"))

prepped_forecast <- format_for_submission(hosp_fitfor$tsfor, method = "ts")
forecast_categorical(prepped_forecast$ensemble, prepped_hosp)
}
}
\references{
\url{https://github.com/cdcepi/Flusight-forecast-data/blob/master/data-experimental/README.md}
}
