% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{forecast_ili}
\alias{forecast_ili}
\title{Forecast ILI}
\usage{
forecast_ili(ilidat, horizon = 4L, trim_date = NULL, constrained = TRUE)
}
\arguments{
\item{ilidat}{Data returned from \link{get_cdc_ili}.}

\item{horizon}{Optional horizon periods through which the forecasts should be generated; default is \code{4}}

\item{trim_date}{Earliest start date you want to use for ILI data. Default \code{NULL} doesn't trim.}

\item{constrained}{Should the model be constrained to a non-seasonal model? Default \code{TRUE} sets PQD(0,0,0) & pdq(0:5,0:5,0:5). See \link[fable:ARIMA]{fable::ARIMA}.}
}
\value{
A named list containing:
\enumerate{
\item \code{ilidat}: The data sent into the function filtered to the location and the \code{trim_date}. Select columns returned.
\item \code{ilidat_tsibble}: The \code{tsibble} class object returned by running \link{make_tsibble} on the data above.
\item \code{ili_fit}: The fit from \link[fabletools:model]{fabletools::model}.
\item \code{ili_forecast}: The forecast from \link[fabletools:forecast]{fabletools::forecast} at the specified horizon.
\item \code{ili_future}: The \code{horizon}-number of weeks of ILI data forecasted into the future.
\item \code{ili_bound}: The data in 1 bound to the data in 5.
\item \code{arima_params}: A tibble with ARIMA model parameters for each location.
\item \code{locstats}: A tibble with missing data information on all locations.
\item \code{removed}: A tibble with locations removed because of high missing ILI data.
}
}
\description{
Forecasts ILI up to specified weeks in the future. Used in downstream modeling.
}
\details{
Currently limited to one location only.
}
\examples{
\dontrun{
# Get data
ilidat <- get_cdc_ili(region=c("national", "state"), years=2010:lubridate::year(lubridate::today()))

# Using data only from march 2020 forward, for US only
ilidat_us <- ilidat \%>\% dplyr::filter(location=="US")
ilifor_us <- forecast_ili(ilidat_us, horizon=4L, trim_date="2020-03-01")
ilifor_us$ili_fit
ilifor_us$arima_params
ilifor_us$ili_forecast
head(ilifor_us$ili_bound)
tail(ilifor_us$ili_bound, 10)
# Plot
library(dplyr)
library(ggplot2)
theme_set(theme_classic())
ilifor_us$ili_bound \%>\%
 mutate(date=cdcfluview::mmwr_week_to_date(epiyear, epiweek)) \%>\%
 filter(date>"2021-03-01") \%>\%
 ggplot(aes(date, ili)) +
 geom_line(lwd=.3, alpha=.5) +
 geom_point(aes(col=forecasted), size=2)

# At the state level
ilifor_st <- forecast_ili(ilidat, horizon=4L, trim_date="2020-03-01")
ilifor_st$ili_fit
ilifor_st$arima_params
ilifor_st$ili_forecast
head(ilifor_us$ili_bound)
tail(ilifor_us$ili_bound, 10)
# Plot
library(dplyr)
library(ggplot2)
theme_set(theme_classic())
ilifor_st$ili_bound \%>\%
  mutate(date=cdcfluview::mmwr_week_to_date(epiyear, epiweek)) \%>\%
  filter(date>"2021-08-01") \%>\%
  ggplot(aes(date, ili, col=forecasted)) +
  geom_line(lwd=.3) +
  geom_point(aes(col=forecasted), size=.7) +
  facet_wrap(~abbreviation, scale="free_y")
}
}
