% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit.R
\name{format_for_submission}
\alias{format_for_submission}
\title{Format forecasts for submission}
\usage{
format_for_submission(.forecasts, method = "ts")
}
\arguments{
\item{.forecasts}{Forecasts to be formatted for submission; if method is "ts" this should be forecasts from \link{ts_fit_forecast}; otherwise this must be a \code{tibble} with forecast output (e.g. output from \link{glm_forecast}) with a colum designating "location"}

\item{method}{Method for forecasting; default is "ts" which will trigger the use of \link{ts_format_for_submission} internally}
}
\value{
A named list of tibbles, one for each model, formatted for submission.
}
\description{
This function prepares influenza hospitalization forecasts in the format required for submission to FluSight.
}
\examples{
\dontrun{
# Get raw data from healthdata.gov
h_raw <- get_hdgov_hosp(limitcols=TRUE)
## save(h_raw, file="~/Downloads/h_raw.rd")
## load(file="~/Downloads/h_raw.rd")

# Prep, and make a tsibble
prepped_hosp <- prep_hdgov_hosp(h_raw, statesonly=TRUE)
prepped_hosp_tsibble <- make_tsibble(prepped_hosp,
                                     epiyear = epiyear,
                                     epiweek=epiweek,
                                     key=location)
# Limit to only Virginia and US
prepped_hosp_tsibble <-
  prepped_hosp_tsibble \%>\%
  dplyr::filter(location \%in\% c("US", "51"))

# Fit a model
hosp_fitfor <- ts_fit_forecast(prepped_hosp_tsibble,
                               horizon=4L,
                               outcome="flu.admits",
                               constrained=TRUE,
                               covariates=c("hosp_rank", "ili_rank"))

# format for submission
formatted_list <- format_for_submission(hosp_fitfor$tsfor, method = "ts")
formatted_list
}
}
\references{
\url{https://github.com/cdcepi/Flusight-forecast-data/blob/master/data-forecasts/README.md}
}
