% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{get_cdc_clin}
\alias{get_cdc_clin}
\title{Retrieve and prep clinical laboratory percent positive flu data}
\usage{
get_cdc_clin(region = "both", years = NULL)
}
\arguments{
\item{region}{Either "state", "national", or "both". Defaults to \code{"both"} to return state and national data combined.}

\item{years}{A vector of years to retrieve data for. CDC has data going back to 1997. Default value (\code{NULL}) retrieves \strong{all} years.}
}
\value{
Tibble with prepared data pulled from NREVSS, including columns for total number of positives, percent positive, epiweek, epiyear
}
\description{
This function returns weekly state and/or national clinical laboratory percent positivity data from the NREVSS reporting instrument.
}
\examples{
\dontrun{
## get all clinical lab flu positivity data
all_clin <- get_cdc_clin()

## or look at a specific location and time
## note: the year "start" will begin at the first epiweek of the season
## this example 2021 will weekly data ...
## ... starting at beginning of 2021/22 season
## ... ending the week before start of 2022/23 season
va_clin <-
  get_cdc_clin(region = "state", years = 2021) \%>\%
  dplyr::filter(location == "51")
}
}
\references{
https://gis.cdc.gov/grasp/fluview/Phase_6_Cleared_Help.pdf
}
