% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{get_cdc_clin}
\alias{get_cdc_clin}
\title{Retrieve clinical laboratory percent positive flu data}
\usage{
get_cdc_clin(region = "both", years = NULL)
}
\arguments{
\item{region}{Either "state", "national", or "both". Defaults to \code{"both"} to return state and national data combined.}

\item{years}{A vector of years to retrieve data for. CDC has data going back to 1997. Default value (\code{NULL}) retrieves \strong{all} years.}
}
\value{
A \code{tibble} with the following columns:
\itemize{
\item \strong{abbreviation}: Abbreviation for the location
\item \strong{location}: FIPS code for the location
\item \strong{epiyear}: Year of reporting (in epidemiological week calendar)
\item \strong{epiweek}: Week of reporting (in epidemiological week calendar)
\item \strong{week_start}: Date of beginning (Sunday) of the given epidemiological week
\item \strong{p_positive}: Percentage of positive specimens
\item \strong{n_positive}: Total number of positive specimens
\item \strong{total}: Total number of specimens tested
}
}
\description{
This function returns weekly state and/or national clinical laboratory percent positivity data from the NREVSS reporting instrument via the CDC FluView API.
}
\examples{
\dontrun{
# Get all clinical lab flu positivity data
all_clin <- get_cdc_clin()
all_clin
# Alternatively look at a specific location and time
# This 2021 will return weekly data
# Starting at beginning of 2021/22 season
# Ending the week before start of 2022/23 season
va_clin <-
  get_cdc_clin(region = "state", years = 2021) \%>\%
  dplyr::filter(location == "51")
va_clin
}
}
\references{
\url{https://gis.cdc.gov/grasp/fluview/Phase_6_Cleared_Help.pdf}
}
