% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{get_cdc_hosp}
\alias{get_cdc_hosp}
\title{Retrieve hospitalization data from FluSurv-NET}
\usage{
get_cdc_hosp(source = "fluview", years = NULL)
}
\arguments{
\item{source}{The source for the hospitalization data; must be one of \code{"fluview"} or \code{"resp-net"}; default is \code{"fluview"}}

\item{years}{A vector of years to retrieve data for (i.e. 2014 for CDC flu season 2014-2015). CDC has data going back to 2009 and up until the \emph{previous} flu season. Default value (\code{NULL}) retrieves \strong{all} years. Only used if \code{source="fluview"}.}
}
\value{
A \code{tibble} with the following columns:
\itemize{
\item \strong{location}: FIPS code for the location
\item \strong{abbreviation}: Abbreviation for the location
\item \strong{region}: Name of the region
\item \strong{epiyear}: Year of reporting (in epidemiological week calendar)
\item \strong{epiweek}: Week of reporting (in epidemiological week calendar)
\item \strong{week_start}: Date of beginning (Sunday) of the given epidemiological week
\item \strong{week_end}: Date of end (Saturday) of the given epidemiological week
\item \strong{rate}: The cumulative rate per 100k
\item \strong{weekly_rate}: The weekly rate per 100k
\item \strong{season}: The flu season to which the given epidemiological week belongs
}
}
\description{
This function retrieves historical FluSurv-NET hospitalization data via either RESP-NET or the CDC FluView API (see 'Details' section).
}
\details{
The data retrieval from FluView and and RESP-NET pulls FluSurv-NET lab-confirmed hospitalization data as overall cumulative rates and weekly incident rates across reporting sites, age groups, sex, and race/ethnicity categories. Note that as of October 2024, the FluSurv-NET hospitalizations from RESP-NET begin in the 2018-19 season, while the FluView data goes back to 2009-10.
}
\examples{
\dontrun{
# Retrieve FluSurv-Net hospitalization data for specific year(s)
get_cdc_hosp()
}
}
\references{
\itemize{
\item \href{https://gis.cdc.gov/GRASP/Fluview/FluHospRates.html}{Hospital Portal}
\item \href{https://github.com/hrbrmstr/cdcfluview}{cdcfluview package}
\item \href{https://data.cdc.gov/Public-Health-Surveillance/Rates-of-Laboratory-Confirmed-RSV-COVID-19-and-Flu/kvib-3txy/about_data}{RESP-NET Data Portal}
\item \href{https://www.cdc.gov/resp-net/dashboard/}{RESP-NET Dashboard}
}
}
