% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{get_cdc_ili}
\alias{get_cdc_ili}
\title{Get ILI data from CDC FluView}
\usage{
get_cdc_ili(region = c("national", "state"), years = NULL)
}
\arguments{
\item{region}{Either "state" or "national" or \code{c("national", "state")} for both.}

\item{years}{A vector of years to retrieve data for. CDC has data going back to 1997. Default value (\code{NULL}) retrieves \strong{all} years.}
}
\value{
A tibble
}
\description{
Get ILI data from CDC FluView
}
\examples{
\dontrun{
get_cdc_ili(region="national", years=2021)
get_cdc_ili(region="state", years=2021) \%>\% dplyr::filter(abbreviation=="VA")
}
}
\references{
cdcfluview documentation: \url{https://hrbrmstr.github.io/cdcfluview/index.html#retrieve-ilinet-surveillance-data}.
}
