% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{get_cdc_vax}
\alias{get_cdc_vax}
\title{Get vaccination data}
\usage{
get_cdc_vax(
  endpoint = "https://data.cdc.gov/resource/k87d-gv3u.json",
  season = NULL,
  limitrows = NULL,
  app_token = Sys.getenv("HEALTHDATA_APP_TOKEN")
)
}
\arguments{
\item{endpoint}{URL to cdc.gov endpoint. See references.}

\item{season}{Limit to a season. Format: 20xx-20yy. E.g., season="2017-2018" or season="2020-2021"}

\item{limitrows}{Limit API query to at most this number of results. Default is \code{NULL} (no limit).}

\item{app_token}{App token from healthdata.gov. If it's \code{NULL} you might get rate limited. Add an entry to your \verb{~/.Renviron} with \code{HEALTHDATA_APP_TOKEN="tokenhere"} that you got from \url{https://healthdata.gov/profile/edit/developer_settings}.}
}
\value{
A tibble
}
\description{
Get vaccination data from cdc.gov endpoint.
}
\examples{
\dontrun{
d <- get_cdc_vax()
d
library(ggplot2)
d \%>\%
  ggplot(aes(date, cumulative_flu_doses)) +
  geom_line() +
  facet_wrap(~season, scale="free_x") +
  theme_bw()
rm(d)
}
}
\references{
API documentation: \url{https://dev.socrata.com/foundry/data.cdc.gov/k87d-gv3u}.
}
