% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{get_hdgov_hosp}
\alias{get_hdgov_hosp}
\title{Get hospitalization data}
\usage{
get_hdgov_hosp(
  endpoint = "https://healthdata.gov/api/views/g62h-syeh/rows.csv",
  app_token = Sys.getenv("HEALTHDATA_APP_TOKEN"),
  limitcols = FALSE
)
}
\arguments{
\item{endpoint}{URL to healthdata.gov endpoint (see references)..}

\item{app_token}{App token from healthdata.gov. If \code{NULL} you might get rate limited. Add an entry to your \verb{~/.Renviron} with \code{HEALTHDATA_APP_TOKEN="tokenhere"} that you got from \url{https://healthdata.gov/profile/edit/developer_settings}.}

\item{limitcols}{Limit the columns returned to the subjectively defined important ones?}
}
\value{
A tibble
}
\description{
Retrieves hospitalization data from the healthdata.gov endpoint with optional filtering on fields, and return the results into a nice tibble.
}
\examples{
\dontrun{
get_hdgov_hosp(limitcols=TRUE)
get_hdgov_hosp(mindate="2021-11-01", limitrows=10, limitcols=FALSE)
get_hdgov_hosp(state="VA")
get_hdgov_hosp(state="VA", mindate="2021-10-01")
get_hdgov_hosp(state="VA", mindate="2021-10-01", maxdate="2021-10-31")
get_hdgov_hosp(state="VA", mindate="2021-10-01", maxdate="2021-11-21", limitrows=5)
}
}
\references{
API documentation: \url{http://dev.socrata.com/foundry/healthdata.gov/g62h-syeh}.
}
