% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{glm_fit}
\alias{glm_fit}
\title{Fit glm models}
\usage{
glm_fit(.data, .models)
}
\arguments{
\item{.data}{Data including all explanatory and outcome variables needed for modeling; must include column for "location"}

\item{.models}{List of models defined as \link[trending]{trending_model} objects}
}
\value{
A \code{tibble} containing characteristics from the "best" \code{glm} model (i.e., the model from ".models" list with lowest RMSE). The columns in this \code{tibble} include:
\itemize{
\item model_class: The "type" of model for the best fit
\item fit: The fitted model object for the best fit
\item location: The geographic
\item data: Original model fit data as a \code{tibble} in a list column
}
}
\description{
This helper function is used in \link[fiphde]{glm_wrap} to fit a list of models and select the best one.
}
\examples{
\dontrun{
##FIXME
}
}
