% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/glm.R
\name{glm_forecast}
\alias{glm_forecast}
\title{Forecast glm models}
\usage{
glm_forecast(
  .data,
  new_covariates = NULL,
  fit,
  alpha = c(0.01, 0.025, seq(0.05, 0.45, by = 0.05)) * 2
)
}
\arguments{
\item{.data}{Data including all explanatory and outcome variables needed for modeling}

\item{new_covariates}{Tibble with one column per covariate, and n rows for n horizons being forecasted}

\item{fit}{Fitted model object from \link[fiphde]{glm_fit}}

\item{alpha}{Vector specifying the threshold(s) to be used for prediction intervals; alpha of \code{0.05} would correspond to 95\% PI; default is \code{c(0.01, 0.025, seq(0.05, 0.45, by = 0.05)) * 2} to range of intervals}
}
\value{
Tibble with forecasts (quantiles and point estimates)
}
\description{
This function uses fitted model object from \link[fiphde]{glm_fit} and future covariate data to create probablistic forecasts at specific quantiles derived from the "alpha" parameter.
}
