% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_forecast}
\alias{plot_forecast}
\title{Plot forecasts}
\usage{
plot_forecast(.data, submission, location = "US", pi = TRUE)
}
\arguments{
\item{.data}{Historical truth data for all locations and outcomes in submission targets}

\item{submission}{Formatted submission}

\item{location}{Vector specifying locations to filter to; 'US' by default.}

\item{pi}{Logical as to whether or not the plot should include 95\% prediction interval; default is \code{TRUE}}
}
\value{
A \code{ggplot2} plot object with line plots for outcome trajectories faceted by location
}
\description{
This function serves as a plotting mechanism for prepped forecast submission data. Using truth data supplied, the plots show the historical trajectory of weekly flu hospitalizations along with the point estimates for forecasts. Optionally, the user can include 95\% prediction interval as well. Plots include trajectories of weekly flu hospitalizations faceted by location.
}
