% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_forecast_categorical}
\alias{plot_forecast_categorical}
\title{fixme}
\usage{
plot_forecast_categorical(categorical_forecast)
}
\arguments{
\item{categorical_forecast}{A forecast from \link{forecast_categorical}}
}
\value{
A plot
}
\description{
fixme
}
\examples{
\dontrun{
h_raw <- get_hdgov_hosp(limitcols=TRUE)
prepped_hosp <- prep_hdgov_hosp(h_raw)
prepped_tsibble <- make_tsibble(prepped_hosp,
                                     epiyear = epiyear,
                                     epiweek=epiweek,
                                     key=location)
# Run with default constrained ARIMA, nonseasonal ETS, no NNETAR
hosp_fitfor <- ts_fit_forecast(prepped_tsibble,
                               horizon=4L,
                               outcome="flu.admits")
prepped_forecast <- format_for_submission(hosp_fitfor$tsfor, method = "ts")
categorical_forecast <- forecast_categorical(prepped_forecast$ensemble, prepped_hosp)
plot_forecast_categorical(categorical_forecast)
}
}
