% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{plot_forecast_categorical}
\alias{plot_forecast_categorical}
\title{Plot categorical forecasts}
\usage{
plot_forecast_categorical(categorical_forecast)
}
\arguments{
\item{categorical_forecast}{A forecast from \link{forecast_categorical}}
}
\value{
A ggplot2 object with categorical forecasts shown as a stacked bar plot.
}
\description{
Creates a bar plot for categorical forecasts. See examples.
}
\examples{
\dontrun{
# Retrieve hospitalization data
h_raw <- get_hdgov_hosp(limitcols=TRUE)
# Prepare and summarize hospitalization data to weekly resolution
prepped_hosp <- prep_hdgov_hosp(h_raw)
# Create a keyed time series tibble with only locations of interest
prepped_tsibble <- make_tsibble(prepped_hosp,
                                     epiyear = epiyear,
                                     epiweek=epiweek,
                                     key=location)
# Run with default constrained ARIMA, nonseasonal ETS, no NNETAR
hosp_fitfor <- ts_fit_forecast(prepped_tsibble,
                               horizon=4L,
                               outcome="flu.admits")
# Prepare forecast for quantile submission format
prepped_forecast <- format_for_submission(hosp_fitfor$tsfor, method = "ts")
# Run categorical summary of quantiles for the time series ensemble
categorical_forecast <- forecast_categorical(prepped_forecast$ensemble, prepped_hosp)
# Plot the categorical forecast
plot_forecast_categorical(categorical_forecast)
}
}
