% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{pois_forc}
\alias{pois_forc}
\title{Simple Poisson count forecaster}
\usage{
pois_forc(.data, .location, .var)
}
\arguments{
\item{.data}{Data frame with incoming data that includes a variable with counts (see ".var" argument), and location (must be stored in a column called "location") and a variable for sorting by date (must be stored in a column called "week_start")}

\item{.location}{The name of the location of interest}

\item{.var}{Bare, unquoted name of the variable with counts to be forecasted}
}
\value{
Vector of length 4 with Poisson forecasts for 4 horizons ahead.
}
\description{
This function is a helper that forecasts Poisson counts for 4 near-term horizons based on characteristics of recently observed count data. The function effectively takes a rolling average of last 4 observations (augmenting with each forecasted horizon as the horizons progress), then uses this average as the parameter for Lambda in a random draw from a Poisson distribution.
}
