% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{replace_ili_nowcast}
\alias{replace_ili_nowcast}
\title{Replace ILINet data with nowcast}
\usage{
replace_ili_nowcast(
  ilidat,
  start_date = NULL,
  weeks_to_replace = 1,
  fallback = TRUE
)
}
\arguments{
\item{ilidat}{ILI data retrieved via \link{get_cdc_ili}}

\item{start_date}{Date from which to start nowcasting; default is \link[lubridate:now]{lubridate::today}}

\item{weeks_to_replace}{Number of weeks of \code{ilidat} to replace; default is \code{2}}

\item{fallback}{Logical as to whether or not to fall back to pseudo nowcast (average of last 4 ILI weeks in the given location) if nowcast data is unavailable; default is \code{TRUE}}
}
\value{
A \code{tibble} with the following columns:
\itemize{
\item \strong{location}: FIPS code for the location
\item \strong{region_type}: The type of location
\item \strong{abbreviation}: Abbreviation for the location
\item \strong{region}: Name of the region
\item \strong{epiyear}: Year of reporting (in epidemiological week calendar)
\item \strong{epiweek}: Week of reporting (in epidemiological week calendar)
\item \strong{week_start}: Date of beginning (Sunday) of the given epidemiological week
\item \strong{weighted_ili}: Population-weighted percentage of ILI outpatient visits
}
}
\description{
This function replaces the weighted ILI retrieved from \link{get_cdc_ili} with nowcast data for each of the locations in the original data. The function will first attempt to use ILI Nearby nowcasts pulled using \link{get_nowcast_ili}. If the ILI Nearby nowcasts are unavailable, the function will optionally fallback to a pseudo nowcast method that averages the observed ILI for the 4 most recent weeks. The nowcast data will be used to add 1 additional week to the observed ILI data and (optionally) replace the number of weeks specified in the "weeks_to_replace" argument.
}
\examples{
\dontrun{
ilidat <- get_cdc_ili(years=2021)
ilidat <-
  ilidat \%>\%
  dplyr::filter(location=="US" | abbreviation=="VA") \%>\%
  dplyr::group_by(location) \%>\%
  dplyr::slice_max(week_start, n=4) \%>\%
  dplyr::select(location:weighted_ili)
ilidat
iliaug <- replace_ili_nowcast(ilidat, weeks_to_replace=1)
iliaug
}
}
