% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wis.R
\name{score_func_param_checker}
\alias{score_func_param_checker}
\title{Score function parameter checker}
\usage{
score_func_param_checker(quantiles, values, actual_value, id = "")
}
\arguments{
\item{quantiles}{vector of forecasted quantiles}

\item{values}{vector of forecasted values}

\item{actual_value}{Actual value.}

\item{id}{name of the score function, e.g. "weighted_interval_score".}
}
\value{
Nothing. Called for side effects.
}
\description{
Unexported utility function to check parameters of a score function. See \link{weighted_interval_score}.
}
