% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{state_replace_ili_nowcast_all}
\alias{state_replace_ili_nowcast_all}
\title{Replace all ILInet data with nowcast data for a state}
\usage{
state_replace_ili_nowcast_all(
  ilidat,
  state,
  impute = TRUE,
  fallback = TRUE,
  ...
)
}
\arguments{
\item{ilidat}{Data from \link{get_cdc_ili}.}

\item{state}{Two-letter state abbreviation to replace completely}

\item{impute}{Logical; try to mean impute missing values using the immediately preceding and following values. See examples.}

\item{fallback}{Logical as to whether or not to fall back to pseudo nowcast (average of last 4 ILI weeks in the given location) if nowcast data is unavailable; default is \code{TRUE}}

\item{...}{Other arguments passed to \link{get_nowcast_ili}, e.g. \code{boundatzero}, which is \code{TRUE} by default.}
}
\value{
The same as the \code{ilidat} input, but with \code{state}'s data from \link{get_cdc_ili} replaced by nowcast data from \link{get_nowcast_ili}.
}
\description{
Replaces \code{weighted_ili} from \link{get_cdc_ili} with nowcast data from \link{get_nowcast_ili} for all dates for a specified location. This is useful for getting data for states where most or all ILI data is missing (e.g., Florida).
}
\details{
This only replaces instances of \code{weighted_ili} in the specified \code{state} where \code{weighted_ili} is \code{NA}. \emph{Most} ILI data from FL is missing, but not all.
}
\examples{
\dontrun{
ilidat <- get_cdc_ili(years=2020)
ilidat <-
  ilidat \%>\%
  dplyr::filter(location=="US" | abbreviation=="VA" | abbreviation=="FL") \%>\%
  dplyr::group_by(location) \%>\%
  dplyr::slice_max(week_start, n=4) \%>\%
  dplyr::select(location:weighted_ili) \%>\%
  dplyr::arrange(location, epiyear, epiweek)
ilidat
state_replace_ili_nowcast_all(ilidat, state="FL")
# Example with Florida, which has a negative value for nowcasted ILI
ilidat <- get_cdc_ili(years=2019)
ilidat <- ilidat \%>\%
  dplyr::filter(location=="US" | abbreviation=="VA" | abbreviation=="FL") \%>\%
  dplyr::filter(epiyear==2020 & epiweek \%in\% c(20, 21, 22)) \%>\%
  dplyr::select(location:weighted_ili) \%>\%
  dplyr::arrange(location, epiyear, epiweek)
ilidat
# defaults to bound at zero
state_replace_ili_nowcast_all(ilidat, state="FL")
# show results when you don't bound at zero
state_replace_ili_nowcast_all(ilidat, state="FL", boundatzero=FALSE)
# example with missing data in florida
ilidat <- get_cdc_ili(region=c("national","state"), years=2019:lubridate::year(lubridate::today()))
ilidat <- ilidat \%>\%
  dplyr::filter(abbreviation=="FL") \%>\%
  dplyr::filter(week_start>="2020-12-13" & week_start<="2021-01-10")
ilidat
state_replace_ili_nowcast_all(ilidat, state="FL")
state_replace_ili_nowcast_all(ilidat, state="FL", impute=FALSE)
}
}
\seealso{
\link{replace_ili_nowcast}
}
