% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{this_monday}
\alias{this_monday}
\title{Get Monday}
\usage{
this_monday()
}
\value{
Date for the Monday of the current week. For more details see \link[lubridate]{floor_date}.
}
\description{
This function is a helper to get the date for the Monday of the current week.
}
