% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/forecast.R
\name{ts_fit_forecast}
\alias{ts_fit_forecast}
\title{Fit and forecast with time-series approaches.}
\usage{
ts_fit_forecast(
  prepped_hosp_tsibble,
  outcome = "flu.admits",
  horizon = 4L,
  trim_date = "2021-10-25",
  constrained = TRUE,
  param_space = list(P = 0, D = 0, Q = 0, p = 1:2, d = 0:2, q = 0),
  covariates = c("hosp_rank", "ili_rank")
)
}
\arguments{
\item{prepped_hosp_tsibble}{fixme}

\item{outcome}{fixme}

\item{horizon}{fixme}

\item{trim_date}{fixme}

\item{constrained}{fixme}

\item{param_space}{fixme}

\item{covariates}{fixme}
}
\value{
A list of the time series fit, forecast, and model formulas
}
\description{
Fit and forecast with time-series approaches.
}
\examples{
\dontrun{
h_raw <- get_hdgov_hosp(limitcols=TRUE)
## save(h_raw, file="~/Downloads/h_raw.rd")
## load(file="~/Downloads/h_raw.rd")
prepped_hosp <- prep_hdgov_hosp(h_raw)
prepped_hosp_tsibble <- make_tsibble(prepped_hosp,
                                     epiyear = epiyear,
                                     epiweek=epiweek,
                                     key=location)
prepped_hosp_tsibble <-
  prepped_hosp_tsibble \%>\%
  dplyr::filter(location \%in\% c("US", "51"))
hosp_fitfor <- ts_fit_forecast(prepped_hosp_tsibble,
                               horizon=4L,
                               outcome="flu.admits",
                               constrained=TRUE,
                               covariates=c("hosp_rank", "ili_rank"))
}
}
