% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/submit.R
\name{ts_format_for_submission}
\alias{ts_format_for_submission}
\title{Format time series forecast}
\usage{
ts_format_for_submission(
  tsfor,
  .target = "wk ahead inc flu hosp",
  .counts = TRUE
)
}
\arguments{
\item{tsfor}{The forecast from \link{ts_fit_forecast}}

\item{.target}{Name of the target in the forecast; default is \code{"wk ahead inc flu hosp"}}

\item{.counts}{Logical; default \code{TRUE} indicates that the target outcome is a count, and should be rounded off at an integer}
}
\value{
A named list of tibbles with probabilistic forecasts (one for each model), formatted for submission with the following columns:
\itemize{
\item \strong{forecast_date}: Date of forecast
\item \strong{target}: Horizon and name of forecasted target
\item \strong{target_end_date}: Last date of the forecasted target (e.g., Saturday of the given epidemiological week)
\item \strong{location}: FIPS code for location
\item \strong{type}: One of either "point" or "quantile" for the forecasted value
\item \strong{quantile}: The quantile for the forecasted value; \code{NA} if "type" is \code{"point"}
\item \strong{value}: The forecasted value
}
}
\description{
This function specifically formats time series forecasts generated with \link{ts_fit_forecast} to adhere to probabilistic forecast submission guidelines for consortia such as FluSight. It is used as a helper in \link{format_for_submission}.
}
\details{
Uses quantiles \code{c(0.01, 0.025, seq(0.05, 0.95, by = 0.05), 0.975, 0.99)} in the built-in \code{fiphde:::q}, using an accessory table \code{fiphde:::quidk}.
}
\examples{
\dontrun{
# Get raw data from healthdata.gov
h_raw <- get_hdgov_hosp(limitcols=TRUE)

# Prep, and make a tsibble
prepped_hosp <- prep_hdgov_hosp(h_raw, statesonly=TRUE)
prepped_hosp_tsibble <- make_tsibble(prepped_hosp,
                                     epiyear = epiyear,
                                     epiweek=epiweek,
                                     key=location)
# Limit to only Virginia and US
prepped_hosp_tsibble <-
  prepped_hosp_tsibble \%>\%
  dplyr::filter(location \%in\% c("US", "51"))

# Fit a model
hosp_fitfor <- ts_fit_forecast(prepped_hosp_tsibble,
                               horizon=4L,
                               outcome="flu.admits",
                               covariates=TRUE)

# Format for submission
formatted_list <- ts_format_for_submission(hosp_fitfor$tsfor)
formatted_list
}
}
\references{
\url{https://github.com/cdcepi/Flusight-forecast-data/blob/master/data-forecasts/README.md}
}
