% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retrieve.R
\name{who_nrevss}
\alias{who_nrevss}
\title{WHO/NREVSS clinical lab surveillance data}
\usage{
who_nrevss(region = c("national", "hhs", "census", "state"), years = NULL)
}
\arguments{
\item{region}{One of "\code{national}", "\code{hhs}", "\code{census}", or "\code{state}"}

\item{years}{A vector of years to retrieve data for (i.e. \code{2014} for CDC
flu season 2014-2015). CDC has data for this API going back to 1997.
Default value (\code{NULL}) means retrieve \strong{all} years. NOTE: if you
specify a 2-digit season value the function will convert that to the corresponding season identifier (i.e. \code{57} == 2017-2018).}
}
\description{
Adapted from \code{cdcfluview::who_nrevss}.

This unexported helper function leverages the CDC FluView API to pull flu surveillance data collected from U.S. World Health Organization (WHO) Collaborating Laboratories and National Respiratory and Enteric Virus Surveillance System (NREVSS) laboratories. The data retrieved can be parameterized by geographic granularity and/or flu season. The function is used internally by \link{get_cdc_clin}.
}
\references{
\itemize{
\item \href{https://github.dev/hrbrmstr/cdcfluview}{cdcfluview package}
}
}
